% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{add_disprop_est}
\alias{add_disprop_est}
\title{Wrapper for implemented disproportionality estimators}
\usage{
add_disprop_est(df, da_estimators)
}
\arguments{
\item{da_estimators}{A character vector containing the desired expected counts.
Defaults to all possible options, i.e. c("rrr", "prr", "ror").}
}
\description{
Wrapper for implemented disproportionality estimators
}
