% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{add_disproportionality}
\alias{add_disproportionality}
\title{Wrapper for adding disproportionality estimates to data frame
containing expected counts}
\usage{
add_disproportionality(
  df,
  da_estimators = c("ic", "prr", "ror"),
  rule_of_N = 3,
  number_of_digits = 2,
  sign_lvl = 0.95
)
}
\arguments{
\item{df}{A data table, or an object possible to convert to a data table, e.g.
a tibble or data.frame. For column specifications, see details.}

\item{da_estimators}{Character vector, defaults to c("ic", "prr", "ror").}

\item{rule_of_N}{Numeric. To protect against spurious
associations due to small observed counts, prr and ror point and
interval estimates are set to NA when the observed is less or equal to
the value passed as 'rule_of_N'. Defaults to 3, but 5 is sometimes used.
Set to NULL if you don't want to apply any such rule.}

\item{number_of_digits}{Set the number of digits to show in output by passing
an integer. Default value is 2 digits. Set to NULL to avoid rounding.}

\item{sign_lvl}{Significance level of confidence or credibility intervals.
Default is 0.95 (i.e. 95 \% confidence interval)}
}
\value{
The passed data frame with additional columns as specified by
parameters.
}
\description{
Wrapper for adding disproportionality estimates to data frame
containing expected counts
}
