% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ci_for_ic}
\alias{ci_for_ic}
\title{Credibility intervals for IC}
\usage{
ci_for_ic(obs, exp, sign_lvl_quantile, shrinkage)
}
\arguments{
\item{obs}{A numeric vector with observed counts, i.e. number of reports
for the selected drug-event-combination.}

\item{exp}{A numeric vector with expected counts, i.e. number of reports
to be expected given a comparator or \emph{background}.}

\item{shrinkage}{A non-negative numeric of length 1, to be added to
observed and expected count. Default is 0.5.}
}
\description{
Produces credibility intervals for Information Component.
}
