% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ci_for_ror}
\alias{ci_for_ror}
\title{Confidence intervals for ROR}
\usage{
ci_for_ror(a, b, c, d, sign_lvl_quantile)
}
\arguments{
\item{a}{Number of reports for the specific drug and event (i.e. the
observed count)}

\item{b}{Number of reports with the drug, without the event}

\item{c}{Number of reports without the drug, with the event}

\item{d}{Number of reports without the drug, without the event}

\item{sign_lvl_quantile}{Argument passed from sign_lvl in \code{\link{ror}}}
}
\description{
Produces (symmetric normality based) confidence intervals
for Reporting Odds Ratio
}
\seealso{
ror
}
