% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{colnames_da}
\alias{colnames_da}
\title{An internal function creating colnames for da confidence/credibility bounds}
\usage{
colnames_da(quantile_prob = list(lower = 0.025, upper = 0.975), da_name = NULL)
}
\arguments{
\item{quantile_prob}{A list with two parameters, lower and upper. Default: list(lower = 0.025, upper = 0.975)}

\item{da_name}{A string, such as "ic", "prr" or "ror". Default: NULL}
}
\value{
A list with two symbols, to be inserted in the dtplyr-chain
}
\description{
Given the output from quantile_prob, and a da_name string,
create column names such as PRR025, ROR025 and IC025
}
