% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{conf_lvl_to_quantile_prob}
\alias{conf_lvl_to_quantile_prob}
\title{Quantile probabilities from confidence level}
\usage{
conf_lvl_to_quantile_prob(conf_lvl = 0.95)
}
\arguments{
\item{conf_lvl}{Confidencelevel of confidence or credibility intervals.
Default is 0.95 (i.e. 95 \% confidence interval)}
}
\value{
A list with two numerical vectors, "lower" and "upper".
}
\description{
Calculates equi-tailed quantile probabilities from a
confidence level
}
\examples{
conf_lvl_to_quantile_prob(0.95)
}
