% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{count_expected}
\alias{count_expected}
\title{Calculate counts required for expected counts, and expected counts}
\usage{
count_expected(dt, da_estimators = c("rrr", "prr", "ror"))
}
\arguments{
\item{dt}{A data table, or an object possible to convert to a data table. See details.}

\item{da_estimators}{A character vector contaning the desired expected counts.
Defaults to c("rrr", "prr", "ror").}
}
\value{
A tibble with counts.
}
\description{
Produces various counts used in disproportionality analysis.
}
\details{
Three columns are expected: "report_id", "drug_name" and "event_name".
Column report_id should be a numeric or characters. The same report_id can
occur on several rows, if the same report contains several drugs and/or
events. Columns drug_name and event_name must be of type character.
}
