% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{count_expected_prr}
\alias{count_expected_prr}
\title{Count expected for Proportional Reporting Rate}
\usage{
count_expected_prr(count_dt)
}
\arguments{
\item{count_dt}{A data table, output from count_expected_rrr}
}
\value{
A data table with added columns for n_event_prr
n_tot_prr and expected_prr
 @export
}
\description{
Internal function to provide expected counts related to the PRR
}
