% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{count_expected_ror}
\alias{count_expected_ror}
\title{Count expected for Reporting Odds Ratio}
\usage{
count_expected_ror(count_dt)
}
\arguments{
\item{count_dt}{A data table, output from count_expected_rrr}
}
\value{
A data table with added columns for n_event_prr,
n_tot_prr and expected_prr

OUTPUT_DESCRIPTION
}
\description{
Internal function to provide expected counts related to the ROR
}
\details{
DETAILS
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}
 \code{\link[tidyselect]{everything}}
}
