% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{count_expected_rrr}
\alias{count_expected_rrr}
\title{Count Expected for Relative Reporting Rate}
\usage{
count_expected_rrr(df, df_colnames, df_syms)
}
\arguments{
\item{df}{See documentation for add_expected_counts}

\item{df_colnames}{See documentation for da}

\item{df_syms}{A list built from df_colnames through conversion to symbols.}
}
\value{
A data frame with columns for obs, n_drug,
n_event, n_tot and (RRR) expected
}
\description{
Internal function to provide expected counts related to the RRR
}
