% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themes.R
\name{custom_colours}
\alias{custom_colours}
\title{Get custom colours}
\usage{
custom_colours(colour = c("blue", "green"))
}
\arguments{
\item{colour}{A character string, must be either "blue" or "green".}
}
\value{
A hex code as a string.
}
\description{
So far, this function provides hex codes for blue and green in
"Uppsala Monitoring Centre - style"
}
\examples{
par(lwd = 5)
hist(runif(30),
  col = pvutils::custom_colours("blue"),
  border = pvutils::custom_colours("green")
)

}
