% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themes.R
\name{custom_ggtheme}
\alias{custom_ggtheme}
\title{Add custom colour theme}
\usage{
custom_ggtheme()
}
\value{
None
}
\description{
This function provides a custom colour ggplot2-theme for blue
and green in "Uppsala Monitoring Centre - style"
}
\examples{
library("ggplot2")
ggplot(mtcars, aes(mpg, disp)) +
  geom_point() +
  pvutils:::custom_ggtheme()

}
