% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{da}
\alias{da}
\title{FUNCTION_TITLE}
\usage{
da(df, ...)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
da_1 <- da(drug_event_df)
}
