% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{grouped_da}
\alias{grouped_da}
\title{Disproportionality Analysis by Subgroups}
\usage{
grouped_da(
  df = NULL,
  df_colnames = NULL,
  df_syms = NULL,
  group_by = NULL,
  expected_count_estimators = NULL,
  da_estimators = NULL,
  sort_by = NULL,
  conf_lvl = NULL,
  rule_of_N = NULL,
  number_of_digits = NULL
)
}
\arguments{
\item{df}{See the da function}

\item{df_colnames}{See the da function}

\item{df_syms}{A list built from df_colnames through conversion to symbols.}

\item{group_by}{For convenience, the df_colnames$group_by is passed as a separate parameter}

\item{expected_count_estimators}{See the da function}

\item{da_estimators}{See the da function}

\item{sort_by}{See the da function}

\item{conf_lvl}{See the da function}

\item{rule_of_N}{See the da function}

\item{number_of_digits}{See the da function}
}
\value{
See the da function
}
\description{
A package internal wrapper for executing da across subgroups
}
\details{
See the da documentation
}
