% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_object_oriented_functions.R
\name{print.da}
\alias{print.da}
\title{print function for da objects}
\usage{
\method{print}{da}(x, n = 10, ...)
}
\arguments{
\item{x}{A S3 obj of class "da", output from \code{pvutils::da()}.}

\item{n}{Control the number of rows to print.}

\item{...}{For passing additional parameters to extended classes.}
}
\value{
Nothing, but prints the tibble da_df in the da object.
}
\description{
print function for da objects
}
\examples{
da_1 <- drug_event_df |> da()
print(da_1)
}
