% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disprop_analysis.R
\name{prr}
\alias{prr}
\title{Proportional Reporting Rate}
\usage{
prr(obs, n_drug, n_event_prr, n_tot_prr, sign_lvl = 0.95)
}
\arguments{
\item{obs}{Number of reports for the specific drug and event (i.e. the
observed count).}

\item{n_drug}{Number of reports with the drug, without the event}

\item{n_event_prr}{Number of reports with the event in the background.}

\item{n_tot_prr}{Number of reports in the background.}

\item{sign_lvl}{Significance level of confidence interval. Default is
0.95 (i.e. 95 \% confidence interval)}
}
\value{
A tibble with three columns (point estimate and credibility bounds).
Number of rows equals length of inputs obs, n_drug, n_event_prr and n_tot_prr.
}
\description{
Calculates Proportional Reporting Rate ("PRR") with
confidence intervals, used in disproportionality analysis.
}
\details{
The PRR is the proportion of reports with an event in set of exposed
cases, divided with the proportion of reports with the event in a background
or comparator, which does not include the exposed.
The \emph{Reporting} highlights the correct interpretation, of examining
reporting as the PRR is calculated from a reporting database.
Note: the function is vectorized, see the examples.
}
\examples{

pvutils::prr(obs = 5, n_drug = 10, n_event_prr = 20, n_tot_prr = 10000)

# Note that input parameters can be vectors (of equal length, no recycling)
pvutils::prr(obs = c(5, 10),
            n_drug = c(10, 20),
            n_event_prr = c(15, 30),
            n_tot_prr = c(10000, 10000))
}
