% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{round_and_sort_by_lower_da_limit}
\alias{round_and_sort_by_lower_da_limit}
\title{Sort a disproportionality analysis by the lower da conf. or cred. limit}
\usage{
round_and_sort_by_lower_da_limit(
  df = NULL,
  df_colnames = NULL,
  df_syms = NULL,
  conf_lvl = NULL,
  sort_by = NULL,
  da_estimators = NULL,
  number_of_digits = 2
)
}
\arguments{
\item{df}{See add_disproportionality}

\item{df_colnames}{See add_disproportionality}

\item{df_syms}{See add_disproportionality}

\item{conf_lvl}{See add_disproportionality}

\item{sort_by}{See add_disproportionality}

\item{da_estimators}{See add_disproportionality}

\item{number_of_digits}{Numeric value. Set the number of digits to show in output by passing
an integer. Default value is 2 digits. Set to NULL to avoid rounding.}
}
\value{
The df object, sorted.
}
\description{
Sorts the output by the mean lower limit of a passed da estimator
}
