% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{round_columns_with_many_decimals}
\alias{round_columns_with_many_decimals}
\title{Rounds columns in da_df with many decimals}
\usage{
round_columns_with_many_decimals(
  da_df = NULL,
  da_estimators = NULL,
  number_of_digits = NULL
)
}
\arguments{
\item{da_df}{See add_disproportionality}

\item{da_estimators}{See add_disproportionality}

\item{number_of_digits}{See add_disproportionality}
}
\value{
A df with rounded columns
}
\description{
Internal function containing a mutate + across
}
