% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{sign_lvl_to_quantile_prob}
\alias{sign_lvl_to_quantile_prob}
\title{Quantile probabilities from significance level}
\usage{
sign_lvl_to_quantile_prob(sign_lvl)
}
\arguments{
\item{sign_lvl}{Significance level of credibility interval. Default is
0.95 (i.e. 95 \% credibility interval)}
}
\value{
A list with two numerical vectors, "lower" and "upper".
}
\description{
Calculates equi-tailed quantile probabilities from a
significance level
}
\examples{
sign_lvl_to_quantile_prob(0.95)
}
