% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{sort_by_lower_da_limit}
\alias{sort_by_lower_da_limit}
\title{Sort a disproportionality analysis by the lower da conf. or cred. limit}
\usage{
sort_by_lower_da_limit(
  df = NULL,
  df_colnames = NULL,
  conf_lvl = NULL,
  sort_by = NULL,
  da_estimators = NULL
)
}
\arguments{
\item{df}{See add_disproportionality}

\item{df_colnames}{See add_disproportionality}

\item{conf_lvl}{See add_disproportionality}

\item{sort_by}{See add_disproportionality}

\item{da_estimators}{See add_disproportionality}
}
\value{
The df object, sorted.
}
\description{
Sorts the output by the mean lower limit of a passed da estimator
}
