% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{summary.da}
\alias{summary.da}
\title{Summary function for disproportionality objects}
\usage{
\method{summary}{da}(object, ...)
}
\arguments{
\item{object}{The object outputted from \code{da}. Extended from the tibble object.}

\item{...}{For passing additional parameters to extended classes.}
}
\value{
Nothing
}
\description{
Provides summary counts of SDRs and shows the top five DECs
}
