% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_excel.R
\name{write_to_excel}
\alias{write_to_excel}
\title{Write to excel}
\usage{
write_to_excel(df, write_path = NULL)
}
\arguments{
\item{df}{The data frame to export. See `?da` for details.}

\item{write_path}{A string giving the file path}
}
\value{
Nothing.
}
\description{
Writes output from a disproportionality analysis to an excel file
}
