% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disprop_analysis.R
\name{add_disprop_est}
\alias{add_disprop_est}
\title{Wrapper for adding disproportionality estimates to data frame
containing expected counts}
\usage{
add_disprop_est(df, da_estimators = c("prr", "ic", "ror"), ...)
}
\arguments{
\item{df}{A data table, or an object possible to convert to a data table, e.g.
a tibble or data.frame. For column specifications, see details.}

\item{da_estimators}{A character vector containing the desired expected counts.
Defaults to all possible options, i.e. c("rrr", "prr", "ror").}

\item{...}{For passing additional arguments, e.g. significance level.}
}
\value{
The passed data frame with additional columns as specified by
parameters.
}
\description{
Wrapper for adding disproportionality estimates to data frame
containing expected counts
}
