% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{add_disproportionality}
\alias{add_disproportionality}
\title{Add disproportionality estimates to data frame
with expected counts}
\usage{
add_disproportionality(
  df = NULL,
  df_colnames = NULL,
  da_estimators = c("ic", "prr", "ror"),
  sort_by = "ror",
  rule_of_N = 3,
  number_of_digits = 2,
  conf_lvl = 0.95
)
}
\arguments{
\item{df}{Intended use is on the output tibble from \code{add_expected_counts}.}

\item{df_colnames}{Provide a list with the column names of the variables to use
passed in \code{df}`i.e. point \code{da} to the column with the report_ids
(\code{report_id}), the drug names (\code{drug}), the adverse event names
(\code{event}) and optionally which subgroups to calculate disproportionality
across through \code{group_by}. See the vignette for examples.}

\item{da_estimators}{Character vector specifying which disproportionality
estimators to use, in case you don't need all implemented options. Defaults
to c("ic", "prr", "ror").}

\item{sort_by}{The output is sorted in descending order of the lower bound of
the confidence/credibility interval for a da estimator, for each drug-event
combination. If drug-event-combination has several groups, the mean is taken
(ignoring NAs). Any of the passed strings in "da_estimators"is accepted,
the default is "ic".}

\item{rule_of_N}{Numeric value. Sets estimates for ROR and PRR to NA when observed
counts are strictly less than the passed value of \code{rule_of_N}. Default value
is 3, 5 is sometimes used as a more liberal alternative. Set to NULL if you
don't want to apply any such rule.}

\item{number_of_digits}{Numeric value. Set the number of digits to show in output by passing
an integer. Default value is 2 digits. Set to NULL to avoid rounding.}

\item{conf_lvl}{Confidencelevel of confidence or credibility intervals.
Default is 0.95 (i.e. 95 \% confidence interval)}
}
\value{
The passed data frame with disproportionality point and interval
estimates.
}
\description{
Add disproportionality estimates to data frame
with expected counts
}
