% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{add_expected_counts}
\alias{add_expected_counts}
\title{Calculate counts required for expected counts, and expected counts}
\usage{
add_expected_counts(df, da_estimators = c("rrr", "prr", "ror"))
}
\arguments{
\item{df}{A data table, or an object possible to convert to a data table, e.g.
a tibble or data.frame. For column specifications, see details.}

\item{da_estimators}{A character vector containing the desired expected counts.
Defaults to all possible options, i.e. c("rrr", "prr", "ror").}
}
\value{
A tibble with counts.
}
\description{
Produces various counts used in disproportionality analysis.
}
\details{
The passed data table should contain three columns: "report_id", "drug_name"
and "event_name". It should have one row per reported drug-event-combination,
i.e. receiving an additional report for drug A and event 1 would add one row
to the table. The same report_id can occur on several rows, if the same
report contains several drugs and/or events. Column report_id must be of type
numeric or character. Columns drug_name and event_name must be
characters.
}
