% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{add_expected_counts}
\alias{add_expected_counts}
\title{Produces expected counts}
\usage{
add_expected_counts(
  df = NULL,
  df_colnames = NULL,
  expected_count_estimators = c("rrr", "prr", "ror")
)
}
\arguments{
\item{df}{An object possible to convert to a data table, e.g.
a tibble or data.frame, containing patient level reported drug-event-pairs.
See below for further details.}

\item{df_colnames}{Provide a list with the column names of the variables to use
passed in \code{df}`i.e. point \code{da} to the column with the report_ids
(\code{report_id}), the drug names (\code{drug}), the adverse event names
(\code{event}) and optionally which subgroups to calculate disproportionality
across through \code{group_by}. See the vignette for examples.}

\item{expected_count_estimators}{A character vector containing the desired
expected count estimators. Defaults to the implemented options, i.e.
c("rrr", "prr", "ror").}
}
\value{
A tibble containing the various counts.
}
\description{
Produces various counts used in disproportionality analysis.
}
\section{The df object}{

 The passed \code{df} should be (convertible to) a data table and at least contain three
 columns: \code{report_id}, \code{drug} and \code{event}. The data table should contain one row
 per reported drug-event-combination, i.e. receiving a single additional report
 for drug X and event Y would add one row to the table. If the single report
 contained drug X for event Y and event Z, two rows would be added, with the
 same \code{report_id} and \code{drug} on both rows. Column \code{report_id} must be of type
 numeric or character. Columns \code{drug} and \code{event} must be of type character.
 You can use other names for your columns, as long as you specify them in df_colnames.
}

