% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_level_disprop_analysis.R
\name{apply_rule_of_N}
\alias{apply_rule_of_N}
\title{apply_rule_of_N}
\usage{
apply_rule_of_N(
  da_df = NULL,
  da_estimators = c("ic", "prr", "ror"),
  rule_of_N = NULL
)
}
\arguments{
\item{da_df}{See the intermediate object da_df in add_disproportionality}

\item{da_estimators}{Default is c("ic", "prr", "ror").}

\item{rule_of_N}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Internal fcn to set da-columns to NA when observed count < 3
}
\details{
Sometimes, you want to protect yourself from spurious findings based
on small observed counts combined with infinitesimal expected counts.
}
