% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{da}
\alias{da}
\title{Disproportionality analysis}
\usage{
da(df, ...)
}
\arguments{
\item{df}{A data table, or an object possible to convert to a data table, e.g.
a tibble or data.frame. For column specifications, see details.}
}
\value{
A tibble with various counts, point and interval estimates for
the ic, prr and ror disproportionality estimators
}
\description{
Execute a disproportionality analysis.
}
\details{
ADD HERE
}
\examples{
da_1 <- da(drug_event_df)
}
