% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_disprop_analysis.R
\name{da}
\alias{da}
\title{Disproportionality Analysis}
\usage{
da(
  df = NULL,
  df_colnames = list(report_id = "report_id", drug = "drug", event = "event", group_by =
    NULL),
  da_estimators = c("ic", "prr", "ror"),
  rule_of_N = 3,
  conf_lvl = 0.95,
  number_of_digits = 2,
  excel_path = NULL
)
}
\arguments{
\item{df}{An object possible to convert to a data table, e.g.
a tibble or data.frame, containing patient level reported drug-event-pairs.
See below for further details.}

\item{df_colnames}{Provide a list with the names of the variables to use in `df`
i.e. report_id, drug, event and optionally group_by.}

\item{da_estimators}{Character vector specifying which disproportionality
estimators to use, in case you don't need all implemented options. Defaults
to c("ic", "prr", "ror").}

\item{rule_of_N}{Numeric value. Sets estimates for ROR and PRR to NA when observed
counts are strictly less than the passed value of \code{rule_of_N}. Default value
is 3, 5 is sometimes used as a more liberal alternative. Set to NULL if you
don't want to apply any such rule.}

\item{conf_lvl}{Confidencelevel of confidence or credibility intervals.
Default is 0.95 (i.e. 95 \% confidence interval)}

\item{number_of_digits}{Numeric value. Set the number of digits to show in output by passing
an integer. Default value is 2 digits. Set to NULL to avoid rounding.}

\item{excel_path}{To write the output of \code{da} to an excel file, provide a path
to a folder e.g. to write to your current working directory, pass \code{getwd()}.
 The excel file will by default be named \code{da.xlsx}. To control the excel file name,
 pass a path ending with the desired filename suffixed with \code{.xlsx}. If you
 do not want to export the output to an excel file, pass NULL (the default).}
}
\value{
\code{da} returns a data frame (invisibly) containing counts and
estimates related to supported disproportionality estimators. Each row
corresponds to a drug-event pair.
}
\description{
\code{da} is used to execute a disproportionality analysis,
potentially across subgroups (see parameter \code{group_by}), using the
Information Component (IC), Proportional Reporting Rate (PRR) and/or the
Reporting Odds Ratio (ROR).
}
\examples{
### Run a disproportionality analysis
da_1 <-
drug_event_df |>
da()

### Run a disproportionality analysis on a data frame with your own column names
# Create the list with colnames for the data
your_own_colnames <-
list(report_id = "repId",
     drug = "drug",
     event = "event",
     group_by = NULL)

# Rename the report_id variable and run again
drug_event_df |>
dplyr::rename(repId = report_id) |>
da(df_colnames = your_own_colnames)

### Run a disproportionality across two subgroups
# Create two groups (even/uneven report_ids) in drug_event_df
drug_event_df_with_grouping  <-
drug_event_df |>
dplyr::mutate("group" = report_id \%\% 2)

# Set the colname for the group_by-parameter to "group"
your_own_colnames <-
     list(report_id = "report_id",
          drug = "drug",
          event = "event",
          group_by = "group")

# Execute da across the subgroups
da_2 <-
drug_event_df_with_grouping |>
da(df_colnames = your_own_colnames)
}
\seealso{
\code{\link[pvutils]{add_expected_counts}}, \code{\link[pvutils]{add_disproportionality}}
}
