% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{Information component}
\usage{
ic(obs, exp, shrinkage = 0.5, sign_lvl = 0.95)
}
\arguments{
\item{obs}{A numeric vector with observed counts, i.e. number of reports
for the selected drug-event-combination.}

\item{exp}{A numeric vector with expected counts, i.e. number of reports
to be expected given a comparator or \emph{background}.}

\item{shrinkage}{A non-negative numeric of length 1, to be added to
observed and expected count. Default is 0.5.}

\item{sign_lvl}{Significance level of credibility interval. Default is
0.95 (i.e. 95 \% credibility interval)}
}
\value{
A tibble with a row for each observed-to-expected pair provided.
}
\description{
Calculates the information component ("IC") and credibility
interval, used in disproportionality analysis.
}
\details{
The IC is based on the relative reporting rate (RRR), but modified with
an addition of "shrinkage" (typically of 0.5) to protect against spurious
associations.

\deqn{\hat{IC} = log_{2}(\frac{\hat{O}+0.5}{\hat{E}+0.5})}

where \eqn{\hat{O}} = observed number of reports, and expected \eqn{\hat{E}}
is (for RRR, and using the entire database as \emph{background}) estimated as

\deqn{ \hat{E} = \frac{N_{drug} \times N_{event}}{N_{TOT}}}

where \eqn{N_{drug}}, \eqn{N_{event}} and \eqn{N_{TOT}} are the number of reports with the drug,
the event, and in the whole database respectively.

From a bayesian perspective, the credibility interval of the IC is constructed
from the poisson-gamma conjugacy. The shrinkage is then a prior distribution of
observed and expected equal to 0.5. The point of \eqn{log_{2}} is to provide
a log-scale for convenient plotting of multiple IC values side-by-side.
}
\examples{
pvutils::ic(obs = 20, exp = 10)

# Note that obs and exp can be vectors (of equal length, no recycling)
pvutils::ic(obs=c(20,30), exp=c(10, 10))

}
