% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_object_oriented_functions.R
\name{summary.da}
\alias{summary.da}
\title{Summary function for disproportionality objects}
\usage{
\method{summary}{da}(object, print = TRUE, ...)
}
\arguments{
\item{object}{A S3 obj of class "da", output from \code{pvutils::da()}.}

\item{print}{Do you want to print the output to the console. Defaults to TRUE.}

\item{...}{For passing additional parameters to extended classes.}
}
\value{
Passes a tibble with the SDR counts invisibly.
}
\description{
Provides summary counts of SDRs and shows the top five DECs
}
