% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_excel.R
\name{write_to_excel}
\alias{write_to_excel}
\title{Calculate counts required for expected counts, and expected counts}
\usage{
write_to_excel(df)
}
\arguments{
\item{df}{A data table, or an object possible to convert to a data table, e.g.}
}
\value{
Nothing.
}
\description{
Produces various counts used in disproportionality analysis.
}
\details{
Placeholder
}
